+++ src/framework/ui-blocker/ui-blocker.js
@@ -27,7 +27,7 @@ export default class UiBlocker {
    * @param {number} config.lowerLimit Время до блокировки интерфейса в миллисекундах. Если вызвать метод unblock раньше, то интерфейс заблокирован не будет
    * @param {number} config.upperLimit Минимальное время блокировки в миллисекундах. Минимальная длительность блокировки
    */
-  constructor({lowerLimit, upperLimit}) {
+  constructor(lowerLimit, upperLimit) {
     this.#lowerLimit = lowerLimit;
     this.#upperLimit = upperLimit;
 
@@ -37,15 +37,15 @@ export default class UiBlocker {
   }
 
   /** Метод для блокировки интерфейса */
-  block() {
+  block = () => {
     this.#startTime = Date.now();
     this.#timerId = setTimeout(() => {
       this.#addClass();
     }, this.#lowerLimit);
-  }
+  };
 
   /** Метод для разблокировки интерфейса */
-  unblock() {
+  unblock = () => {
     this.#endTime = Date.now();
     const duration = this.#endTime - this.#startTime;
 
@@ -60,7 +60,7 @@ export default class UiBlocker {
     }
 
     setTimeout(this.#removeClass, this.#upperLimit - duration);
-  }
+  };
 
   /** Метод, добавляющий CSS-класс элементу */
   #addClass = () => {
