--- src/mock/point-types.js
@@ -1,9 +1,0 @@
-const POINT_TYPES = [
-  {
-    id: 1,
-    type: 'bus',
-    label: 'label'
-  }
-];
-
-export const getPointTypes = () => POINT_TYPES;
