+++ src/mock/type-point.js
@@ -6,4 +6,4 @@ const POINT_TYPES = [
   }
 ];
 
-export const getpointType = () => POINT_TYPES;
+export const getPointTypes = () => POINT_TYPES;
