+++ src/util.js
@@ -10,5 +10,7 @@ const getRandomInteger = (a = 0, b = 1) => {
 const humanizeTime = (dueDate) => dayjs(dueDate).format('HH:mm');
 const humanizeDate = (dueDate) => dayjs(dueDate).format('DD MMM');
 const humanizeDateTime = (dueDate) => dayjs(dueDate).format('YY/MM/DD HH:mm');
+const humanizeDateTimeForFilters = (dueDate) => dayjs(dueDate).format('YYYY-MM-DD');
 
-export {getRandomInteger, humanizeTime, humanizeDate, humanizeDateTime};
+
+export {getRandomInteger, humanizeTime, humanizeDate, humanizeDateTime, humanizeDateTimeForFilters};
