--- src/view/trip-list.js
@@ -1,9 +1,0 @@
-import AbstractView from '../framework/view/abstract-view.js';
-
-const createTripListTemplate = () => '<ul class="trip-events__list"></ul>';
-
-export default class TripList extends AbstractView{
-  get template() {
-    return createTripListTemplate();
-  }
-}
