+++ src/view/view-eventlist.js
@@ -1,23 +1,9 @@
-import {createElement} from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
 const createTripListTemplate = () => '<ul class="trip-events__list"></ul>';
 
-export default class viewList {
-  #element = null;
-
+export default class ViewList extends AbstractView{
   get template() {
     return createTripListTemplate();
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
