+++ src/view/view-filter.js
@@ -1,37 +1,36 @@
-import {createElement} from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
-const createFiltersTemplate = () => (`
-<form class="trip-filters" action="#" method="get">
-<div class="trip-filters__filter">
-  <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything">
-  <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
-</div>
+const createFiltersTemplate = (filters) => {
+  // eslint-disable-next-line
+  const futureFilter = filters || '';
 
-<div class="trip-filters__filter">
-  <input id="filter-future" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="future">
-  <label class="trip-filters__filter-label" for="filter-future">Future</label>
-</div>
+  return (
+    `
+    <form class="trip-filters" action="#" method="get">
 
-<button class="visually-hidden" type="submit">Accept filter</button>
-</form>
-`);
+      <div class="trip-filters__filter">
+        <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything">
+        <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
+      </div>
 
-export default class FiltersView {
-  #element = null;
+      <div class="trip-filters__filter">
+        <input id="filter-future" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="future">
+        <label class="trip-filters__filter-label" for="filter-future">Future</label>
+      </div>
+
+      <button class="visually-hidden" type="submit">Accept filter</button>
+
+    </form>
+    `);
+};
+
+export default class ViewFilter extends AbstractView {
+  constructor(filters) {
+    super();
+    this.filters = filters;
+  }
 
   get template() {
-    return createFiltersTemplate();
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
-  removeElement() {
-    this.#element = null;
+    return createFiltersTemplate(this.filters);
   }
 }
